backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
duckdb
google-generativeai>=0.7.2
numpy
pandas
fpdf2>=2.7.9
python-dotenv>=1.0.1
aiofiles>=23.2.1
